requirements.txt
numpy==1.22.3
torch==1.11.0
biopython
matplotlib